split storage 1 with tools
