+++ tkcore/cpp_lib/CMakeLists.txt
@@ -10,15 +10,15 @@ add_library(lib_tkcore
         ${CMAKE_CURRENT_LIST_DIR}/find/key_storage_v2.cpp
 
         # storage 1
-
+        ${CMAKE_CURRENT_LIST_DIR}/storage1/pass_spliter_v1.cpp
+        ${CMAKE_CURRENT_LIST_DIR}/storage1/salt_test1.cpp
+        ${CMAKE_CURRENT_LIST_DIR}/storage1/key_storage_v1.cpp
 
         # storage 2
-        ${CMAKE_CURRENT_LIST_DIR}/storage1/pass_spliter_v1.cpp
+
         ${CMAKE_CURRENT_LIST_DIR}/salt_text/salt2.cpp
         ${CMAKE_CURRENT_LIST_DIR}/salt_text/salt2_schema.cpp
-        ${CMAKE_CURRENT_LIST_DIR}/salt_text/salt_test1.cpp
         ${CMAKE_CURRENT_LIST_DIR}/salt_text/salt_base.cpp
-        ${CMAKE_CURRENT_LIST_DIR}/storage1/key_storage_v1.cpp
         ${CMAKE_CURRENT_LIST_DIR}/main/storage2/split_password.cpp
         ${CMAKE_CURRENT_LIST_DIR}/main/storage2/storage.cpp
         ${CMAKE_CURRENT_LIST_DIR}/main/storage2/storage_structure.cpp
