+++ tkcore/cpp_lib/storage1/key_storage_v1.h
@@ -7,7 +7,7 @@
 
 #include "core/key_core.h"
 #include "list"
-#include "salt_text/salt_test1.h"
+#include "salt_test1.h"
 
 namespace thekey_v1 {
