+++ tkcore/cpp_lib/storage1/salt_test1.cpp
@@ -7,7 +7,7 @@
 //
 
 #include "salt_test1.h"
-#include "salt_base.h"
+#include "salt_text/salt_base.h"
 #include <string.h>
 #include <cstring>
 #include <limits.h>
